cloud_infra/mqtt_subscriber/requirements.txt
paho-mqtt==1.6.1
grpcio==1.60.0
joblib==1.3.2
protobuf==4.21.12
threadpoolctl==3.2.0
